Replaces simple mock files by inline mocks for better readability
Tests are easier to read if all mocked information is in place instead of having it in separate files. On the other hand, long mocked inputs and outputs can make the test lengthy and unreadable. This replaces all mocked files that contain objects with just one key with inline objects. Longer mock files are kept.